Repairing an inconsistency in a malloc.

+++ sphere.c
@@ -117,7 +117,7 @@ g_printerr ("npcs=%d\n", d->sphere.npcs);
 
     vectori_realloc (&d->sphere.pcvars, d->sphere.npcs);
 
-    dtmp = (gdouble *) g_malloc0 (d->nrows * sizeof (gfloat));
+    dtmp = (gdouble *) g_malloc0 (d->nrows * sizeof (gdouble));
     for (j=0; j<d->sphere.npcs; j++) {
       vname = g_strdup_printf ("PC%d", j+1);
       newvar_add_with_values (dtmp, d->nrows, vname, d, gg);
